+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -85,6 +85,12 @@ New checks
 
   Ensures lambda expressions' parameter lists are explicit.
 
+- New :doc:`bsl-namespace-global
+  <clang-tidy/checks/bsl-namespace-global>` check.
+
+  Warns if anything other than main(), operator new/delete, namespaces, or
+  extern "C" declarations are present at global scope.
+
 - New :doc:`bsl-op-bitwise-operands
   <clang-tidy/checks/bsl-op-bitwise-operands>` check.